microservices/flask-frontend/requirements.txt
Flask==2.3.2
authlib==1.3.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
